+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -276,8 +276,12 @@ export interface CanvasKit {
      *              the image is destroyed.
      * @param info - If provided, will be used to determine the width/height/format of the
      *               source image. If not, sensible defaults will be used.
+     * @param srcIsPremul - set to true if the src data has premultiplied alpha. Otherwise, it will
+     *         be assumed to be Unpremultiplied. Note: if this is true and info specifies
+     *         Unpremul, Skia will not convert the src pixels first.
      */
-    MakeLazyImageFromTextureSource(src: TextureSource, info?: ImageInfo | PartialImageInfo): Image;
+    MakeLazyImageFromTextureSource(src: TextureSource, info?: ImageInfo | PartialImageInfo,
+                                   srcIsPremul?: boolean): Image;
 
     /**
      * Deletes the associated WebGLContext. Function not available on the CPU version.
@@ -1487,7 +1491,7 @@ export interface Canvas extends EmbindObject<Canvas> {
      *          not supported in JS, so that colorType corresponds to raw bytes Uint8Array.
      */
     readPixels(srcX: number, srcY: number, imageInfo: ImageInfo, dest?: MallocObj,
-               bytesPerRow?: number): Uint8Array | Float32Array | null;
+               bytesPerRow?: number): Float32Array | Uint8Array | null;
 
     /**
      * Removes changes to the current matrix and clip since Canvas state was
@@ -2665,8 +2669,12 @@ export interface Surface extends EmbindObject<Surface> {
      * @param src
      * @param info - If provided, will be used to determine the width/height/format of the
      *               source image. If not, sensible defaults will be used.
+     * @param srcIsPremul - set to true if the src data has premultiplied alpha. Otherwise, it will
+     *               be assumed to be Unpremultiplied. Note: if this is true and info specifies
+     *               Unpremul, Skia will not convert the src pixels first.
      */
-    makeImageFromTextureSource(src: TextureSource, info?: ImageInfo | PartialImageInfo): Image | null;
+    makeImageFromTextureSource(src: TextureSource, info?: ImageInfo | PartialImageInfo,
+                               srcIsPremul?: boolean): Image | null;
 
     /**
      * Returns current contents of the surface as an Image. This image will be optimized to be
@@ -2715,8 +2723,11 @@ export interface Surface extends EmbindObject<Surface> {
      *
      * @param img - A texture-backed Image.
      * @param src - A valid texture source of any dimensions.
+     * @param srcIsPremul - set to true if the src data has premultiplied alpha. Otherwise, it will
+     *               be assumed to be Unpremultiplied. Note: if this is true and the image was
+     *               created with Unpremul, Skia will not convert.
      */
-    updateTextureFromSource(img: Image, src: TextureSource): void;
+    updateTextureFromSource(img: Image, src: TextureSource, srcIsPremul?: boolean): void;
 
     /**
      * Returns the width of this surface in pixels.